pyproject.toml
[project]
name = "vuln-research-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "0din-jef>=0.1.6",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
